Change CalculateNetworkFee to be public
